fix: move code back from base_utils to utils path
