+++ test-orchestrator/test_orchestrator/api/asset.py
@@ -37,7 +37,7 @@ from typing import Dict
 from fastapi import APIRouter, Depends
 
 from test_orchestrator.auth import verify_auth
-from test_orchestrator.base_utils import submodel_validation
+from test_orchestrator.utils import submodel_validation
 from test_orchestrator.errors import Error, HTTPError
 from test_orchestrator.utils.special_characteristics import (
     process_notification_and_retrieve_dtr
