+++ test-orchestrator/test_orchestrator/utils/product_carbon_footprint.py
@@ -32,7 +32,7 @@ from test_orchestrator.request_handler import make_request
 from test_orchestrator.auth import get_dt_pull_service_headers
 from test_orchestrator.errors import Error, HTTPError
 from test_orchestrator.cache import CacheProvider
-from test_orchestrator.base_utils import get_dataplane_access
+from test_orchestrator.utils import get_dataplane_access
 
 logger = logging.getLogger(__name__)
